+++ internal/controller/grafanaorganization_controller.go
@@ -36,7 +36,7 @@ type GrafanaOrganizationReconciler struct {
 }
 
 func SetupGrafanaOrganizationReconciler(mgr manager.Manager, conf config.Config) error {
-	grafanaAPI, err := grafanaclient.GenerateGrafanaClient(conf.GrafanaURL, conf)
+	grafanaAPI, err := grafanaclient.GenerateGrafanaClient(conf.GrafanaURL, conf, "organization-controller")
 	if err != nil {
 		return fmt.Errorf("unable to create grafana client: %w", err)
 	}